+++ .env
@@ -1,0 +1,2 @@
+GMAILUSER="oyamakayano@gmail.com"
+GMAILPASSWORD="lopnylgqbfnjnqjr"
