+++ package.json
@@ -14,6 +14,7 @@
     "autoprefixer": "10.4.14",
     "formik": "^2.4.2",
     "next": "13.4.7",
+    "nodemailer": "^6.9.3",
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
